Removed rng from color_pulse triggers
- The position for the color pulses was *always* random as written. If you wanted to pulse in specific locations consistently, it was impossible. To fix this, I am changing the intended behavior of color pulses to expect the starting_offset parameter to be populated by the programming calling the `lighting_controller::trigger()` function.

+++ src/utility.rs
@@ -159,7 +159,7 @@ impl<'a> StatefulRainbow<'a> {
     }
 
     pub fn current_color(&self) -> RGBA8 {
-        self.backer[self.position.get_current() as usize]
+        self.backer[self.position.get_current()]
     }
 
     pub fn decrement(&mut self) {
@@ -171,11 +171,11 @@ impl<'a> StatefulRainbow<'a> {
     }
 
     pub fn peek_next_color(&self) -> RGBA8 {
-        self.backer[self.position.peek_next() as usize]
+        self.backer[self.position.peek_next()]
     }
 
     pub fn peek_last_color(&self) -> RGBA8 {
-        self.backer[self.position.peek_prev() as usize]
+        self.backer[self.position.peek_prev()]
     }
 
     pub fn reset(&mut self) {
